+++ cpmpy/expressions/variables.py
@@ -201,7 +201,7 @@ def cpm_array(arr):
     """
     if not isinstance(arr, np.ndarray):
         arr = np.array(arr)
-    return NDVarArray(shape=arr.shape, dtype=type(arr.flat[0]), buffer=arr)
+    return NDVarArray(shape=arr.shape, dtype=arr.dtype, buffer=arr)
 
 
 class NullShapeError(Exception):
@@ -241,6 +241,11 @@ class _NumVarImpl(Expression):
     # for sets/dicts. Because names are unique, so is the str repr
     def __hash__(self):
         return hash(str(self))
+
+    def deepcopy(self, memodict={}):
+        copied = type(self)(self.lb, self.ub, self.name)
+        copied._value = self.value()
+        return copied
 
 class _IntVarImpl(_NumVarImpl):
     """
@@ -350,6 +355,9 @@ class NegBoolView(_BoolVarImpl):
     def __invert__(self):
         return self._bv
 
+    def deepcopy(self, memodict={}):
+        return NegBoolView(self._bv.deepcopy(memodict))
+
 
 # subclass numericexpression for operators (first), ndarray for all the rest
 class NDVarArray(Expression, np.ndarray):
@@ -371,6 +379,10 @@ class NDVarArray(Expression, np.ndarray):
 
     def value(self):
         return np.reshape([x.value() for x in self], self.shape)
+
+    def deepcopy(self, memodict={}):
+        copied = [arg.deepcopy(memodict) if isinstance(arg, Expression) else arg for arg in self]
+        return cpm_array(copied)
     
     def __repr__(self):
         """