+++ cpmpy/solvers/__init__.py
@@ -13,6 +13,8 @@
 
         ortools
         pysat
+        gurobi
+        pysdd
         utils
 
     ===============
@@ -23,6 +25,8 @@
 
         CPM_ortools
         CPM_pysat
+        CPM_gurobi
+        CPM_pysdd
 
     =================
     List of functions
@@ -37,3 +41,5 @@ from .utils import builtin_solvers, get_supported_solvers, param_combinations
 from .ortools import CPM_ortools
 from .pysat import CPM_pysat
 from .minizinc import CPM_minizinc
+from .gurobi import  CPM_gurobi
+from .pysdd import CPM_pysdd